+++ app/minime/models.py
@@ -5,7 +5,7 @@ class Url(models.Model):
     hash = models.SlugField(max_length=6)
     url = models.URLField(max_length=2048)
     created = models.DateTimeField(auto_now=True)
-    count = models.IntegerField(default=0)
+    password = models.CharField(max_length=1024, null=True)
 
     def __str__(self):
         return self.url