+++ HISTORY.md
@@ -3,7 +3,8 @@
 
 ### Next
 
-- Feature: Added Django Rest Framework.
+- Added password protection.
+- Added Django Rest Framework.
 - Added test coverage and coveralls.
 - Added redis cache.
 - Added ability to run app on Heroku.